rbd-target-api: get image's exclusive lock owner and break it
Once the client logged in and IOs are done, one of the iscsi
gateway nodes will hold the exclusive lock and be the lock
owner in tcmu-runner.

When we try rollback the snap shot of the image, it will become
readonly and failed even DELETE the LUN from the client host.

This patch will just get the lock owner and break it when doing
the LUN's DELETE opration.

Signed-off-by: Zhang Zhuoyu <zhangzhuoyu@cmss.chinamobile.com>
Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>